move lang/fungus.* to src
