+++ build.zig
@@ -25,9 +25,7 @@ pub fn build(b: *std.build.Builder) anyerror!void {
             "-Wextra",
             "-Wpedantic",
             "-Wvla",
-            "-std=c11",
-
-            "-Isrc"
+            "-std=c11"
         });
 
         if (mode == std.builtin.Mode.Debug) {
@@ -43,6 +41,7 @@ pub fn build(b: *std.build.Builder) anyerror!void {
 
         const sources = [_][]const u8 {
             "main.c",
+            "fungus.c",
 
             // lexical analysis
             "lex.c",
@@ -53,7 +52,6 @@ pub fn build(b: *std.build.Builder) anyerror!void {
             "lang/rules.c",
             "lang/precedence.c",
             "lang/pattern.c",
-            "lang/fungus.c",
             "lang/ast_expr.c",
 
             // sema
