--- src/fun_types.h
@@ -1,13 +1,0 @@
-#ifndef FUN_TYPES_H
-#define FUN_TYPES_H
-
-#include <stddef.h>
-#include <stdint.h>
-
-typedef uint32_t hsize_t;
-
-typedef _Bool           fun_bool;
-typedef long long int   fun_int;
-typedef long double     fun_float;
-
-#endif
