+++ src/fungus.c
